Individual Assignments/A4/assignment-4/partB/requirements.txt
google-cloud-storage
boto3~=1.24.25
botocore~=1.27.25
Levenshtein~=0.18.2